Lift the typing indicator out of the input component
